Fixed the code
Tested the code and implemented new function for category, restaurant and user

+++ Online-Food-Ordering/src/main/java/com/atm/model/Order.java
@@ -48,7 +48,7 @@ public class Order {
 	
 	private int totalItem;
 	
-	private int totalPrice;
+	private Long totalPrice;
 	
 //	private Payment payment;
